Refactor calendar.js into sidebar folder
Inline calendar into the sidebar folder.
